requirements.txt
imutils==0.5.4
keras==2.9.0
Keras-Preprocessing==1.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.3
numpy==1.22.4
oauthlib==3.2.0
opencv-contrib-python==4.6.0.66
opencv-contrib-python-headless==4.7.0.68
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.68
pandas==1.4.2
Pillow==9.1.1
scikit-image==0.19.3
scikit-learn==1.2.1
scikit-plot==0.3.7
scipy==1.10.0
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0